backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-environ
drf-yasg
django-filter
django-widget-tweaks

# ve varsa diğer paketleriniz…
